requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
Appium-Python-Client==1.2.0
attrs==21.2.0
BeautifulReport==0.1.3
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
Flask==2.0.1
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
PyYAML==5.4.1
requests==2.26.0
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
